Incorporating feedback from PR#1

+++ src/interfaces/_external/IChainlinkAggregatorV3.sol
@@ -7,5 +7,11 @@ interface IChainlinkAggregatorV3 {
     function latestRoundData()
         external
         view
-        returns (uint80 roundId, int256 answer, uint256 startedAt, uint256 updatedAt, uint80 answeredInRound);
+        returns (
+            uint80 roundId,
+            int answer,
+            uint startedAt,
+            uint updatedAt,
+            uint80 answeredInRound
+        );
 }
